🏗️ Refactored the action menu
