+++ src/client/components/ComicDetail/ComicVineMatchPanel.tsx
@@ -2,22 +2,26 @@ import React, { ReactElement } from "react";
 import { ComicVineSearchForm } from "../ComicVineSearchForm";
 import MatchResult from "./MatchResult";
 import { isEmpty } from "lodash";
+import { useStore } from "../../store";
+import { useShallow } from "zustand/react/shallow";
 
 export const ComicVineMatchPanel = (comicVineData): ReactElement => {
-  const {
-    comicObjectId,
-    comicVineSearchQueryObject,
-    comicVineAPICallProgress,
-    comicVineSearchResults,
-  } = comicVineData.props;
+  const { comicObjectId, comicVineMatches } = comicVineData.props;
+  const { comicvine } = useStore(
+    useShallow((state) => ({
+      comicvine: state.comicvine,
+    })),
+  );
   return (
     <>
-      <div className="search-results-container">
-        {!isEmpty(comicVineSearchResults) && (
+      <div>
+        {!isEmpty(comicVineMatches) ? (
           <MatchResult
-            matchData={comicVineSearchResults}
+            matchData={comicVineMatches}
             comicObjectId={comicObjectId}
           />
+        ) : (
+          <>{comicvine.scrapingStatus}</>
         )}
       </div>
     </>
