+++ src/client/components/ComicDetail/MatchResult.tsx
@@ -1,8 +1,9 @@
-import React, { useCallback } from "react";
+import React from "react";
 import { isNil, map } from "lodash";
-import { applyComicVineMatch } from "../../actions/comicinfo.actions";
 import { convert } from "html-to-text";
 import ellipsize from "ellipsize";
+import { LIBRARY_SERVICE_BASE_URI } from "../../constants/endpoints";
+import axios from "axios";
 
 interface MatchResultProps {
   matchData: any;
@@ -14,12 +15,16 @@ const handleBrokenImage = (e) => {
 };
 
 export const MatchResult = (props: MatchResultProps) => {
-  const applyCVMatch = useCallback(
-    // (match, comicObjectId) => {
-    //   dispatch(applyComicVineMatch(match, comicObjectId));
-    // },
-    [],
-  );
+  const applyCVMatch = async (match, comicObjectId) => {
+    return await axios.request({
+      url: `${LIBRARY_SERVICE_BASE_URI}/applyComicVineMetadata`,
+      method: "POST",
+      data: {
+        match,
+        comicObjectId,
+      },
+    });
+  };
   return (
     <>
       {map(props.matchData, (match, idx) => {
@@ -32,45 +37,29 @@ export const MatchResult = (props: MatchResultProps) => {
           });
         }
         return (
-          <div className="search-result mb-4" key={idx}>
-            <div className="columns">
-              <div className="column is-one-fifth">
+          <div className="mb-4" key={idx}>
+            <div className="flex flex-row">
+              <div className="w-full mr-2">
                 <img
-                  className="cover-image"
+                  className="rounded-md"
                   src={match.image.thumb_url}
                   onError={handleBrokenImage}
                 />
               </div>
-
-              <div className="search-result-details column">
-                <div className="is-size-5">{match.name}</div>
-
-                <div className="field is-grouped is-grouped-multiline mt-1">
-                  <div className="control">
-                    <div className="tags has-addons">
-                      <span className="tag">Number</span>
-                      <span className="tag is-primary">
-                        {match.issue_number}
-                      </span>
-                    </div>
-                  </div>
-                  <div className="control">
-                    <div className="tags has-addons">
-                      <span className="tag">Cover Date</span>
-                      <span className="tag is-warning">{match.cover_date}</span>
-                    </div>
-                  </div>
-                </div>
-                <div className="is-size-7">
+              <div>
+                <div className="">{match.name}</div>
+                <span>Number</span>
+                <span>{match.issue_number}</span>
+                <span className="tag">Cover Date</span>
+                <span className="tag is-warning">{match.cover_date}</span>
+                <div className="text-sm">
                   {ellipsize(issueDescription, 300)}
                 </div>
               </div>
             </div>
 
-            <div className="vertical-line"></div>
-
-            <div className="columns ml-6 volume-information">
-              <div className="column is-one-fifth">
+            <div className="">
+              <div className="">
                 <img
                   src={match.volumeInformation.results.image.icon_url}
                   className="cover-image"
@@ -78,42 +67,30 @@ export const MatchResult = (props: MatchResultProps) => {
                 />
               </div>
 
-              <div className="column">
-                <div className="is-size-6">{match.volume.name}</div>
-                <div className="field is-grouped is-grouped-multiline mt-2">
-                  <div className="control">
-                    <div className="tags has-addons">
-                      <span className="tag">Total Issues</span>
-                      <span className="tag is-warning">
-                        {match.volumeInformation.results.count_of_issues}
-                      </span>
-                    </div>
-                  </div>
-
-                  <div className="control">
-                    <div className="tags has-addons">
-                      <span className="tag">Publisher</span>
-                      <span className="tag is-warning">
-                        {match.volumeInformation.results.publisher.name}
-                      </span>
-                    </div>
-                  </div>
+              <div className="">
+                <div className="">{match.volume.name}</div>
+                <div className="">
+                  <span className="">Total Issues</span>
+                  <span className="">
+                    {match.volumeInformation.results.count_of_issues}
+                  </span>
                 </div>
               </div>
+
+              <span className="tag">Publisher</span>
+              <span className="tag is-warning">
+                {match.volumeInformation.results.publisher.name}
+              </span>
             </div>
-            <div className="columns">
-              <div className="column">
-                <button
-                  className="button is-normal is-outlined is-primary is-light is-pulled-right"
-                  onClick={() => applyCVMatch(match, props.comicObjectId)}
-                >
-                  <span className="icon is-size-5">
-                    <i className="fas fa-clipboard-check"></i>
-                  </span>
-                  <span>Apply Match</span>
-                </button>
-              </div>
-            </div>
+            <button
+              className="button is-normal is-outlined is-primary is-light is-pulled-right"
+              onClick={() => applyCVMatch(match, props.comicObjectId)}
+            >
+              <span className="icon is-size-5">
+                <i className="fas fa-clipboard-check"></i>
+              </span>
+              <span>Apply Match</span>
+            </button>
           </div>
         );
       })}
